feat: `let`/`mut` compound assignments
This change allows compound assignment operators to be used with `let`
and `mut` variable declarations. For example, `let x += 1` is now valid
and is treated as equivalent to `let x = $x + 1`. As such, a parser
error is thrown if `$x` is not already defined when `let x += 1` is
parsed. This shorthand is particularly useful within nested scopes and
closures.

Other changes include:

Allow variables such as `$$env` and `$$$in` (i.e. a reserved/builtin
variable name prefixed with two or more `$`) to be defined. The reason
for this is consistency, as e.g. `$$foo` is currently both allowed and
considered to be a different variable than `$foo` by the engine.
Declaring a variable named `$env` etc. is still disallowed.

`let` and `mut` calls now utilize the same code for parsing, enhancing
consistency and maintainability.

Change `nu-engine::compile::operator::decompose_assignment()` to a
public method of `Assignment`.

Add examples and update command info of `let` and `mut`.