Implement per-pipeline process groups for background jobs
Currently in nushell, when running in an interactive session,
every pipeline gets its own process group, which is set
as the foreground process group. Meanwhile, in non-interactive sessions,
all processes use the shell's process group.

This commit makes so that background jobs also use one processs group
per pipeline, but without moving them to foreground.

This commit also fixes a bug in nushell's trunk, which occurs if
a process in a non-interactive session gets signaled with the TSTP
signal, which causes the shell to try to regain foreground status, and
read a confirmation message from stdin (which should not occur in an
interactive session).

The `foreground.rs`'s inner `foreground_pgrp` module has been renamed to
`child_pgrp` since it now manages process groups in general.